+++ packages/core/src/components/icon-button/icon-button.interfaces.ts
@@ -1,0 +1,5 @@
+export type IconButtonVariant = "secondary" | "tertiary" | "map";
+
+export interface IconButtonClickEvent {
+  originalEvent: MouseEvent;
+}