+++ storybook/cypress/e2e/panel.cy.ts
@@ -28,8 +28,9 @@ describe("Panel", () => {
     cy.get("dso-panel.hydrated")
       .invoke("prop", "closeButtonLabel", "Labeltest")
       .shadow()
-      .contains("button", "Labeltest")
-      .should("exist");
+      .find("dso-icon-button")
+      .shadow()
+      .find("button[aria-label='Labeltest']");
   });
 
   it("is emphasized", () => {